add logger @Slf4j

+++ src/main/java/com/example/demo/services/impl/UserServiceImpl.java
@@ -8,6 +8,8 @@ import com.example.demo.model.dto.UserUpdate;
 import com.example.demo.model.entities.UserEntity;
 import com.example.demo.repositories.UserRepository;
 import com.example.demo.services.UserService;
+import lombok.extern.slf4j.Slf4j;
+import org.hibernate.ObjectNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,14 +18,44 @@ import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
+import static java.lang.String.format;
+
+
 @Service
+@Slf4j
 public class UserServiceImpl implements UserService {
+
+    private static final String NO_USERS_FOUND_WARN = "no users found in DB";
+
+    private static final String AGE_VALID_ERROR = "user with %s id age validation error";
+
+    private static final String USER_DELETE_INFO = "user with %s id has been deleted";
+
+    private static final String USER_DELETE_ERROR = "user with the given ID does not exist and can't be deleted";
+
+    private static final String USER_UPDATE_ERROR = "user with given id does not exist and can't be updated";
+
+    private static final String GET_USER_BY_ID_ERROR = "user with given id doesn't exist";
 
     private final UserRepository userRepository;
 
     private final UserMapper userMapper;
 
     private final UserEntityMapper userEntityMapper;
+
+    private void ageValidation (int age, long id) {
+        if (age < 0) {
+            log.error(format(AGE_VALID_ERROR, id));
+            throw new RuntimeException(format(AGE_VALID_ERROR, id));
+        }
+    }
+
+    private void ageValidation (int age) {
+        if (age < 0) {
+            log.error(AGE_VALID_ERROR);
+            throw new RuntimeException(AGE_VALID_ERROR);
+        }
+    }
 
     public UserServiceImpl(final UserRepository userRepository, final UserMapper userMapper, final UserEntityMapper userEntityMapper) {
         this.userRepository = userRepository;
@@ -33,11 +65,15 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public List<User> getUsers() {
-        return userRepository
+        final List<User> users = userRepository
                 .findAll()
                 .stream()
                 .map(userMapper::map)
                 .toList();
+        if (users.isEmpty()) {
+            log.warn(NO_USERS_FOUND_WARN);
+        }
+        return users;
     }
 
     @Override
@@ -47,24 +83,46 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User getUserById(final long id) {
-        return userMapper.map(userRepository.getById(id));
+        Optional<UserEntity> userEntity = userRepository.findById(id);
+        if (userEntity.isEmpty()) {
+            log.error(GET_USER_BY_ID_ERROR);
+            throw new RuntimeException(GET_USER_BY_ID_ERROR);
+        }
+
+        return userMapper.map(userEntity.get());
     }
 
     @Override
     public void addUser(final UserCreation user) {
+        ageValidation(user.getAge());
         userRepository.save(userEntityMapper.map(user));
     }
 
     @Override
     public void updateUser(final UserUpdate userUpdate) {
-        final LocalDateTime creationDate = userRepository.getById(userUpdate.getId()).getCreatedAt();
-        final UserEntity updatedUserEntity = userEntityMapper.map(userUpdate);
-        updatedUserEntity.setCreatedAt(creationDate);
+        ageValidation(userUpdate.getAge(), userUpdate.getId());
+        Optional<UserEntity> userEntity = userRepository.findById(userUpdate.getId());
+
+        if (userEntity.isEmpty()) {
+            log.error(USER_UPDATE_ERROR);
+            throw new RuntimeException(USER_UPDATE_ERROR);
+        }
+
+        final UserEntity updatedUserEntity = userEntity.get();
+        updatedUserEntity.setName(userUpdate.getName());
+        updatedUserEntity.setAge(userUpdate.getAge());
+        updatedUserEntity.setRole(userUpdate.getRole());
+        updatedUserEntity.setPassword(userUpdate.getPassword());
         userRepository.save(updatedUserEntity);
     }
 
     @Override
     public void deleteUser(final long id) {
+        if (!userRepository.existsById(id)) {
+            log.error(USER_DELETE_ERROR);
+            throw new RuntimeException(USER_DELETE_ERROR);
+        }
         userRepository.deleteById(id);
+        log.info(format(USER_DELETE_INFO, id));
     }
 }
